Prevent certain explosions from destroying item frames/paintings
Currently, any sort of player- or entity-initiated explosions such as creepers, withers and TNT will destroy item frames and paintings when detonated. This commit prevents this behavior for mob explosions.

Due to Bukkit API limitations, it is hard to differentiate between a player hitting a frame and a player igniting TNT which then explodes the frame. This means that this fix does not cover player-ignited TNT, as it will still destroy the hanging entity.